ugly-pet-comfyui/requirements.txt
comfyui>=2.4.0
torch>=2.1.0
safetensors
Pillow
numpy
